+++ src/index.js
@@ -153,6 +153,28 @@ export const set = curry(
 );
 
 /**
+ * @function transform
+ *
+ * @description
+ * perform same operation as set, but using a callback function that receives
+ * the value (and additional parameters, if provided) to get the value to set
+ *
+ * @param {Array<number|string>|number|string} path the path to set the value at
+ * @param {function} fn the function to transform the retrieved value with
+ * @param {Array<*>|Object} object the object to set the value in
+ * @param {...Array<any>} extraArgs additional arguments to pass to the transform function
+ * @returns {Array<*>|Object} a new object with the same structure and the value assigned
+ */
+export const transform = curry(
+  (path, fn, object, ...extraArgs) =>
+    isEmptyPath(path)
+      ? fn(object, ...extraArgs)
+      : getDeepClone(path, object, (ref, key) => (ref[key] = fn(ref[key], ...extraArgs))),
+  // eslint-disable-next-line no-magic-numbers
+  3
+);
+
+/**
  * @function add
  *
  * @description